Cleanup the cache tree only if directory exists
This is useful when running the test locally (multiple times).
But if the directory does not exist, it does not need to be removed

+++ unittests/test_directory_cache.py
@@ -29,7 +29,8 @@ class Test(parser_test_case.parser_test_case_t):
         """
 
         # Clear the cache tree
-        shutil.rmtree(self.cache_dir)
+        if os.path.isdir(self.cache_dir):
+            shutil.rmtree(self.cache_dir)
 
         # Test with compression OFF
         cache = parser.directory_cache_t(directory=self.cache_dir)
